splitting main.go into a few files for clarity
